Adds a limit of 4 hours to water tracker
Also adds a thank you command. '!thanks'

The limit was put in place so that users won't get messages if they
forget to use the command '!water stop'. This is for the case where a
user might go to sleep after activating it and wake up with 8 messages
or more.